fix: move voodoo stuff into its own folder
