+++ src/voodoo/index.ts
@@ -1,0 +1,1 @@
+export { default as VoodooClient } from './VoodooClient'
